Split Proxy and Memento lecture, improved Memento examples
